added paths to footer buttons and component structure
